requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.7.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
scipy>=1.9.0
scikit-learn>=1.1.0
